pyproject.toml
[tool.poetry]
name = "local-chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
chroma = "^0.2.0"
langchain = "^0.0.340"
logger = "^1.4"
click = "^8.1.7"
rich = "^13.7.0"
bs4 = "^0.0.1"
openai = "^1.3.5"
chromadb = "^0.4.18"
tiktoken = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
